home can call Api and fetch data

+++ app/src/main/java/com/iti/elfarsisy/mad41/myapplication/data/model/Current.kt
@@ -17,7 +17,7 @@ data class Current(
     val visibility: Int? = null,
 
     @field:SerializedName("uvi")
-    val uvi: Int? = null,
+    val uvi: Double? = null,
 
     @field:SerializedName("pressure")
     val pressure: Int? = null,
